Show SELinux mode in device info always
This check was implemented at the time, because compiling with SELinux (HAVE_SELINUX=true) was still an option. Since Android 4.3 SELinux became enabled by default, HAVE_SELINUX guards (basically, if-commands) were removed and somewhere in the process the PROPERTY_OVERRIDE "ro.build.selinux" was lost. Instead of adding it in like other ROMs, I found the cause of the problem and fix it here.

More info:
https://android-review.googlesource.com/#/c/38683/
https://android-review.googlesource.com/#/c/44822/
http://review.cyanogenmod.org/#/c/47003/1